refactor: getTalkerId middleware
